Se ha agregado un nuevo campo en la entidad Horario y se ha realizado la migración respectiva, de igual forma la vista se crea desde "código"

+++ HorariosConsoleApp/Entities/DetalleHorario.cs
@@ -2,7 +2,7 @@
 {
     public class ConsultaDetalleHoras
     {
-        public decimal NumeroHoras { get; set; }
+        public int NumeroHoras { get; set; }
         public int TipoHoraId { get; set; }
         public string TipoHora { get; set; }
         public string Horario { get; set; }
